Use searchguard-ssl for providing transport TLS
This fixes access privileges that were unnecessary after porting to
using the DNS discovery approach. It also makes sure we do not mount too
many volumes when there are no extra plugins to install.
For the main change it uses the internal CA for Kubernetes to provision
signed certificates and an initContainer that uses this API to generate
a key and a certificate on a shared volume. It requires a FQDN for the
discovery service, hence the new clusterDomain key.